requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.3.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.6
django-cors-headers==4.3.0
django-crontab==0.7.1
django-dbbackup==4.0.2
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
cloudinary==1.36.0